Fix another cgo pointer check in xpc_darwin
This is another place that causes the error:
panic: runtime error: cgo argument has Go pointer to Go pointer

Reference the rules here: https://github.com/golang/go/issues/12416

Passing a pointer to a slice is seen as passing a Go pointer to C where the Go memory the pointer points to contains another Go pointer. Slices fall into this category. Passing a pointer to the first element in a slice is fair game, however.